+++ colonyShip.go
@@ -14,9 +14,9 @@ func newColonyShip() *colonyShip {
 	s.BaseCargoCapacity = 7500
 	s.BaseSpeed = 2500
 	s.FuelConsumption = 1000
-	s.RapidfireFrom = map[ID]int{DeathstarID: 250}
-	s.RapidfireAgainst = map[ID]int{EspionageProbeID: 5, SolarSatelliteID: 5}
+	s.RapidfireFrom = map[ID]int64{DeathstarID: 250}
+	s.RapidfireAgainst = map[ID]int64{EspionageProbeID: 5, SolarSatelliteID: 5}
 	s.Price = Resources{Metal: 10000, Crystal: 20000, Deuterium: 10000}
-	s.Requirements = map[ID]int{ShipyardID: 4, ImpulseDriveID: 3}
+	s.Requirements = map[ID]int64{ShipyardID: 4, ImpulseDriveID: 3}
 	return s
 }
