+++ crystalMine.go
@@ -16,14 +16,14 @@ func newCrystalMine() *crystalMine {
 }
 
 // EnergyConsumption returns the building energy consumption
-func (b *crystalMine) EnergyConsumption(level int) int {
-	return int(math.Ceil(10 * float64(level) * math.Pow(1.1, float64(level))))
+func (b *crystalMine) EnergyConsumption(level int64) int64 {
+	return int64(math.Ceil(10 * float64(level) * math.Pow(1.1, float64(level))))
 }
 
 // Production returns the crystal production of the mine
-func (b *crystalMine) Production(universeSpeed int, productionRatio, globalRatio float64, plasmaTech, level int) int {
+func (b *crystalMine) Production(universeSpeed int64, productionRatio, globalRatio float64, plasmaTech, level int64) int64 {
 	basicIncome := 15.0
 	levelProduction := 20 * float64(universeSpeed) * (1 + float64(plasmaTech)*0.0066) * float64(level) * math.Pow(1.1, float64(level))
-	production := int(levelProduction*productionRatio*globalRatio + (basicIncome * float64(universeSpeed)))
+	production := int64(levelProduction*productionRatio*globalRatio + (basicIncome * float64(universeSpeed)))
 	return production
 }
