+++ heavyLaser.go
@@ -12,7 +12,7 @@ func newHeavyLaser() *heavyLaser {
 	d.StructuralIntegrity = 8000
 	d.ShieldPower = 100
 	d.WeaponPower = 250
-	d.RapidfireFrom = map[ID]int{BomberID: 10, DeathstarID: 100}
-	d.Requirements = map[ID]int{ShipyardID: 4, EnergyTechnologyID: 3, LaserTechnologyID: 6}
+	d.RapidfireFrom = map[ID]int64{BomberID: 10, DeathstarID: 100}
+	d.Requirements = map[ID]int64{ShipyardID: 4, EnergyTechnologyID: 3, LaserTechnologyID: 6}
 	return d
 }
