+++ shieldingTechnology.go
@@ -10,6 +10,6 @@ func newShieldingTechnology() *shieldingTechnology {
 	b.ID = ShieldingTechnologyID
 	b.IncreaseFactor = 2.0
 	b.BaseCost = Resources{Metal: 200, Crystal: 600}
-	b.Requirements = map[ID]int{ResearchLabID: 6, EnergyTechnologyID: 3}
+	b.Requirements = map[ID]int64{ResearchLabID: 6, EnergyTechnologyID: 3}
 	return b
 }
