+++ smallCargo.go
@@ -14,9 +14,9 @@ func newSmallCargo() *smallCargo {
 	s.BaseCargoCapacity = 5000
 	s.BaseSpeed = 5000
 	s.FuelConsumption = 10
-	s.RapidfireFrom = map[ID]int{BattlecruiserID: 3, HeavyFighterID: 3, DeathstarID: 250}
-	s.RapidfireAgainst = map[ID]int{EspionageProbeID: 5, SolarSatelliteID: 5}
+	s.RapidfireFrom = map[ID]int64{BattlecruiserID: 3, HeavyFighterID: 3, DeathstarID: 250}
+	s.RapidfireAgainst = map[ID]int64{EspionageProbeID: 5, SolarSatelliteID: 5}
 	s.Price = Resources{Metal: 2000, Crystal: 2000}
-	s.Requirements = map[ID]int{ShipyardID: 2, CombustionDriveID: 2}
+	s.Requirements = map[ID]int64{ShipyardID: 2, CombustionDriveID: 2}
 	return s
 }
